Re-ify the concept of TIME-FORMAT
Up to this point, we had a very informal notion of what a time
format value actually might be. Now, there is a type for it, which
records not only the generated printer function, but also the
source from which the function was generated. This hopefully will
allow us to also generate parsers using the same source expression.